+++ src/particles/WeatherGenerator.hh
@@ -1,4 +1,4 @@
-//  $Id: WeatherGenerator.hh,v 1.5 2001/04/03 10:45:51 grumbel Exp $
+//  $Id: WeatherGenerator.hh,v 1.6 2001/08/13 21:35:38 grumbel Exp $
 // 
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 2000 Ingo Ruhnke <grumbel@gmx.de>
@@ -35,7 +35,7 @@ public:
   ///
   virtual void update(float delta) =0;
   ///
-  int get_z_pos() const { return 1000; }
+  float get_z_pos() const { return 1000; }
   ///
   void draw_offset(int, int, float) {}
